requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
sqlalchemy
mysql-connector-python
openpyxl
notebook
jupyter
